Add project_loader and data_router now use project_loader
