+++ rasa_nlu/data_router.py
@@ -27,6 +27,7 @@ from twisted.internet import reactor
 from twisted.internet.defer import Deferred
 from twisted.logger import jsonFileLogObserver, Logger
 from typing import Text, Dict, Any, Optional, List
+from rasa_nlu.project_loader import ProjectLoader
 
 logger = logging.getLogger(__name__)
 
@@ -105,7 +106,9 @@ class DataRouter(object):
         else:
             self.component_builder = ComponentBuilder(use_cache=True)
 
-        self.project_store = self._create_project_store(project_dir)
+        self.project_loader = ProjectLoader(project_dir,
+                                            remote_storage,
+                                            component_builder)
         self.pool = ProcessPool(self._training_processes)
 
     def __del__(self):
@@ -143,47 +146,6 @@ class DataRouter(object):
                         "(No 'request_log' directory configured)")
             return None
 
-    def _collect_projects(self, project_dir):
-        if project_dir and os.path.isdir(project_dir):
-            projects = os.listdir(project_dir)
-        else:
-            projects = []
-
-        projects.extend(self._list_projects_in_cloud())
-        return projects
-
-    def _create_project_store(self, project_dir):
-        projects = self._collect_projects(project_dir)
-
-        project_store = {}
-
-        for project in projects:
-            project_store[project] = Project(self.component_builder,
-                                             project,
-                                             self.project_dir,
-                                             self.remote_storage)
-
-        if not project_store:
-            default_model = RasaNLUModelConfig.DEFAULT_PROJECT_NAME
-            project_store[default_model] = Project(
-                    project_dir=self.project_dir,
-                    remote_storage=self.remote_storage)
-        return project_store
-
-    def _list_projects_in_cloud(self):
-        try:
-            from rasa_nlu.persistor import get_persistor
-            p = get_persistor(self.remote_storage)
-            if p is not None:
-                return p.list_projects()
-            else:
-                return []
-        except Exception:
-            logger.exception("Failed to list projects. Make sure you have "
-                             "correctly configured your cloud storage "
-                             "settings.")
-            return []
-
     @staticmethod
     def _create_emulator(mode):
         """Create emulator for specified mode.
@@ -209,32 +171,15 @@ class DataRouter(object):
         return self.emulator.normalise_request_json(data)
 
     def parse(self, data):
-        project = data.get("project", RasaNLUModelConfig.DEFAULT_PROJECT_NAME)
+        project = data.get("project")
         model = data.get("model")
 
-        if project not in self.project_store:
-            projects = self._list_projects(self.project_dir)
-
-            cloud_provided_projects = self._list_projects_in_cloud()
-            projects.extend(cloud_provided_projects)
-
-            if project not in projects:
-                raise InvalidProjectError(
-                    "No project found with name '{}'.".format(project))
-            else:
-                try:
-                    self.project_store[project] = Project(
-                            self.component_builder, project,
-                            self.project_dir, self.remote_storage)
-                except Exception as e:
-                    raise InvalidProjectError(
-                        "Unable to load project '{}'. Error: {}".format(
-                            project, e))
+        project_instance = self.project_loader.load_project(project)
 
         time = data.get('time')
-        response, used_model = self.project_store[project].parse(data['text'],
-                                                                 time,
-                                                                 model)
+        response, used_model = project_instance.parse(data['text'],
+                                                      time,
+                                                      model)
 
         if self.responses:
             self.responses.info('', user_input=response, project=project,
@@ -242,22 +187,18 @@ class DataRouter(object):
 
         return self.format_response(response)
 
-    @staticmethod
-    def _list_projects(path):
-        """List the projects in the path, ignoring hidden directories."""
-        return [os.path.basename(fn)
-                for fn in utils.list_subdirectories(path)]
-
     def parse_training_examples(self, examples, project, model):
         # type: (Optional[List[Message]], Text, Text) -> List[Dict[Text, Text]]
         """Parses a list of training examples to the project interpreter"""
 
+        project_instance = self.project_loader.load_project(project)
+
         predictions = []
         for ex in examples:
             logger.debug("Going to parse: {}".format(ex.as_dict()))
-            response, _ = self.project_store[project].parse(ex.text,
-                                                            None,
-                                                            model)
+            response, _ = project_instance.parse(ex.text,
+                                                 None,
+                                                 model)
             logger.debug("Received response: {}".format(response))
             predictions.append(response)
 
@@ -274,7 +215,7 @@ class DataRouter(object):
         return {
             "available_projects": {
                 name: project.as_dict()
-                for name, project in self.project_store.items()
+                for name, project in self.project_loader.get_projects().items()
             }
         }
 
@@ -285,27 +226,26 @@ class DataRouter(object):
         if not project:
             raise InvalidProjectError("Missing project name to train")
 
-        if project in self.project_store:
-            if self.project_store[project].status == 1:
-                raise AlreadyTrainingError
-            else:
-                self.project_store[project].status = 1
-        elif project not in self.project_store:
-            self.project_store[project] = Project(
-                    self.component_builder, project,
-                    self.project_dir, self.remote_storage)
-            self.project_store[project].status = 1
+        project_instance = self.project_loader.load_project(project)
+
+        if project_instance.status == 1:
+            raise AlreadyTrainingError
+        else:
+            project_instance.status = 1
 
         def training_callback(model_path):
             model_dir = os.path.basename(os.path.normpath(model_path))
-            self.project_store[project].update(model_dir)
+            project_instance.update(model_dir)
             return model_dir
 
         def training_errback(failure):
-            logger.warn(failure)
-            target_project = self.project_store.get(
-                failure.value.failed_target_project)
-            if target_project:
+            logger.warning(failure)
+            try:
+                target_project = self.project_loader.load_project(
+                    failure.value.failed_target_project)
+            except InvalidProjectError:
+                pass
+            else:
                 target_project.status = 0
             return failure
 
@@ -330,10 +270,6 @@ class DataRouter(object):
         model = model or None
         file_name = utils.create_temporary_file(data, "_training_data")
         test_data = load_data(file_name)
-
-        if project not in self.project_store:
-            raise InvalidProjectError("Project {} could not "
-                                      "be found".format(project))
 
         preds_json = self.parse_training_examples(test_data.intent_examples,
                                                   project,
@@ -371,12 +307,11 @@ class DataRouter(object):
 
         if project is None:
             raise InvalidProjectError("No project specified".format(project))
-        elif project not in self.project_store:
-            raise InvalidProjectError("Project {} could not "
-                                      "be found".format(project))
+
+        project_instance = self.project_loader.load_project(project)
 
         try:
-            unloaded_model = self.project_store[project].unload(model)
+            unloaded_model = project_instance.unload(model)
             return unloaded_model
         except KeyError:
             raise InvalidProjectError("Failed to unload model {} "
